bot/requirements.txt
pandas
sentence-transformers
faiss-cpu
numpy
llama-index
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-vector-stores-faiss